style(draco-core): clear pre-existing clippy lints
Auto-fixable idiom lints (is_multiple_of, collapsible if, redundant casts, div_ceil, needless cfg, etc.) plus manual fixes: combine identical match arms, let-else to ?, range loop instead of manual counter, struct-init, an is_empty method, and #[allow(type_complexity)] on inherently-complex codec signatures. Behavior-preserving; full draco-core + draco-io + draco-gltf test suites pass.

+++ crates/draco-core/src/mesh_edgebreaker_traversal_valence_decoder.rs
@@ -270,7 +270,7 @@ impl<'a> EdgebreakerTraversalDecoder for MeshEdgebreakerTraversalValenceDecoder<
         } else {
             active_valence
         };
-        self.active_context = (clamped - self.min_valence) as i32;
+        self.active_context = clamped - self.min_valence;
 
         // Record processed connectivity corner (like InternalTraversalDecoder)
         self.processed_connectivity_corners.push(corner.0);
